hide experimental sections from frontpage nav if not root user
