+++ apps/hub/src/components/FrontpageNav/index.tsx
@@ -1,0 +1,61 @@
+import { FC, Suspense } from "react";
+
+import {
+  PageMenu,
+  PageMenuLink,
+  PageMenuSubheader,
+} from "@/components/ui/PageMenu";
+import {
+  definitionsRoute,
+  epistemicAgentsRoute,
+  evaluationsRoute,
+  groupsRoute,
+  questionSetsRoute,
+  variablesRoute,
+} from "@/lib/routes";
+import { auth } from "@/lib/server/auth";
+import { isAdminUser } from "@/users/auth";
+
+import { FrontpageNavDesktopHeader } from "./FrontpageNavDesktopHeader";
+import { MobileFrontpageNavHeaderText } from "./MobileFrontpageNavHeaderText";
+
+const MobileFrontpageNavHeader: FC = () => {
+  return (
+    <div className="text-sm font-medium text-slate-600">
+      <MobileFrontpageNavHeaderText />
+    </div>
+  );
+};
+
+const FrontpageNavEvalsIfRoot: FC = async () => {
+  const session = await auth();
+  if (!session || !isAdminUser(session.user)) {
+    return null;
+  }
+
+  return (
+    <>
+      <PageMenuSubheader>Experimental</PageMenuSubheader>
+      <PageMenuLink name="Question Sets" href={questionSetsRoute()} />
+      <PageMenuLink name="Evals" href={evaluationsRoute()} />
+      <PageMenuLink name="Epistemic Agents" href={epistemicAgentsRoute()} />
+    </>
+  );
+};
+
+export const FrontpageNav: FC = () => {
+  return (
+    <PageMenu
+      mobileHeader={<MobileFrontpageNavHeader />}
+      desktopHeader={<FrontpageNavDesktopHeader />}
+    >
+      <PageMenuLink name="Models" href="/" />
+      <PageMenuLink name="Variables" href={variablesRoute()} />
+      <PageMenuLink name="Definitions" href={definitionsRoute()} />
+      <PageMenuLink name="Groups" href={groupsRoute()} />
+      <Suspense>
+        <FrontpageNavEvalsIfRoot />
+      </Suspense>
+    </PageMenu>
+  );
+};
